requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.34
PyMySQL==1.1.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
